Componentization and folder optimization
