--- src/App.js
@@ -1,10 +1,0 @@
-import React from "react";
-import UserPage from "./components/users";
-
-export default function App() {
-  return (
-    <div className="App">
-      <UserPage />
-    </div>
-  );
-}
