+++ src/components/App.js
@@ -1,0 +1,10 @@
+import React from "react";
+import UserPage from "./users";
+
+export default function App() {
+    return (
+        <div className="App">
+            <UserPage/>
+        </div>
+    );
+}
