+++ src/components/users/index.jsx
@@ -1,0 +1,15 @@
+import AddUser from "./AddUser";
+import TableUser from "../layouts/table/Table";
+import {useState} from "react";
+
+export default function UsersPage() {
+    const [users, setUsers] = useState([]);
+
+    return (
+        <>
+            <AddUser addUser={setUsers}/>
+            
+            <TableUser users={users}/>
+        </>
+    );
+}
